refactor code structure to resuse java api, separate submodule for antrun tests, encode string to avoid token error, abstact from oracle as much as possible
