+++ .gitignore
@@ -1,0 +1,5 @@
+target
+pom.bck
+.classpath
+.project
+.settings
